+++ lib/foldable_runtime.dart
@@ -13,9 +13,11 @@ export 'src/model/display_info.dart';
 export 'src/model/fold_posture.dart';
 export 'src/model/foldable_state.dart';
 export 'src/model/hinge.dart';
-export 'src/model/posture_resolver.dart' show FoldingFeatureState, PostureResolver;
+export 'src/model/posture_resolver.dart'
+    show FoldingFeatureState, PostureResolver;
 export 'src/model/posture_thresholds.dart';
 export 'src/model/quirks.dart';
+export 'src/platform/dart_plugin_registrant.dart';
 export 'src/platform/foldable_runtime_platform.dart';
 export 'src/platform/unsupported_foldable_runtime.dart';
 export 'src/widgets/foldable_builder.dart';